[cl/rope] Do the GPU RoPE rotation in fp32 -> fixes qwen3 sampling degeneration
The GPU RoPE kernel (rope_inplace_f16) read the cos/sin LUT and the Q/K
activations as half AND did the rotation `lo*c - hi*s` / `hi*c + lo*s`
entirely in half. For head_dim=128 with a large rope_theta (qwen3) that
fp16 multiply-add distorts the rotated Q/K just enough that the attention
softmax DISTRIBUTION is wrong while the argmax survives: greedy decoding
stays coherent (only the top token matters) but SAMPLING (do_sample=true)
walks the distorted tail and degenerates into loops ("a story and a
story", "\boxed{}") that compound autoregressively.

This was mis-diagnosed earlier as a d128 *flash* bug and worked around by
forcing qwen3 onto the host prefill path (min_prefill_thr head_dim<256 ->
UINT_MAX), which cost ~16x prefill (122 vs ~2300 TPS). The flash kernel
was fine; the RoPE was the culprit. Root fix: cast the operands to float,
rotate in fp32, store fp16 (the activation dtype is unchanged -- only the
intermediate math is fp32). Then revert the min_prefill_thr host gate so
qwen3 uses GPU flash + GPU RoPE like every other model.

Verified Intel Xe3, qwen3_lg_q6k, do_sample=true (temp 0.6/top_k 20):
1K continuation coherent (was degenerate), prefill ~2255-2534 TPS (was
122 on the host gate), greedy short answer "Seoul". gemma4 "Seoul" and
gemma2 coherent -- no regression (fp32 rotation applies to them too).

+++ nntrainer/tensor/cl_operations/attention_kernels.cpp
@@ -797,12 +797,19 @@ __kernel void rope_inplace_f16(__global const half *in,
   if (t >= M || h >= num_heads || k >= half_d) return;
   long row = (long)t * num_heads * (2 * half_d) + (long)h * (2 * half_d);
   long lut = (long)(start_pos + t) * half_d + k;
-  half c = cos_lut[lut];
-  half s = sin_lut[lut];
-  half lo = in[row + k];
-  half hi = in[row + k + half_d];
-  out[write_off + row + k]          = lo * c - hi * s;
-  out[write_off + row + k + half_d] = hi * c + lo * s;
+  // Rotate in FP32. Activations (Q/K) are FP16 by design, but doing the
+  // cos/sin multiply-add in half loses enough precision that the rotated Q/K
+  // are subtly wrong -- the ATTENTION SCORE argmax survives (greedy stays
+  // coherent) but the softmax DISTRIBUTION is distorted, which SAMPLING
+  // exposes (qwen3 head_dim=128 / high rope_theta degenerates under
+  // do_sample=true). fp32 rotation makes the distribution match the host RoPE;
+  // only the final store is FP16 (the activation dtype).
+  float c = (float)cos_lut[lut];
+  float s = (float)sin_lut[lut];
+  float lo = (float)in[row + k];
+  float hi = (float)in[row + k + half_d];
+  out[write_off + row + k]          = (half)(lo * c - hi * s);
+  out[write_off + row + k + half_d] = (half)(hi * c + lo * s);
 }
 __kernel void scatter_copy_f16(__global const half *in, __global half *out,
                                const int N) {
